[MachineLICM] Let targets decide if copy-like instructions are cheap
When checking whether it is profitable to hoist an instruction, the pass
may override a target's ruling because it assumes that all COPY
instructions are cheap, and that may not be the case for all
micro-architectures.

On AArch64 there's 0% difference in performance in LLVM's test-suite.
Additionally, very few tests were affected by this change which shows
how useful it is to keep it.